Require Freereg1CsvFile model from processor library
I added code to manually load the Freereg1CsvFile model from the FreeregCsvProcessor
library code which depends on it.  I'm not quite sure this is the correct thing to do
as it may circumvent the rails dependency management and require restarts of the
rails server in development mode whenever we've changed the model file.