Correct 2 minor processing issues and improved method names
Corrected situation where the Character set field was inferred as
present but was actually blank.
Corrected processing of Register name field when it was empty. In this
situation the code had used the Place name as the Register name. It now
produces the correct blank field.
Took the opportunity to change some method names to improve readability.