Master Place Names
Changes have been made to use a Master Place Name Collection and to
manage that collection on-line, to use the lat/lon from there in the
Place documents together with the Genuki url. In addition we now use the
uploaded date from the tar file as the last modified date for a file and
its records. Creation of the search records docs is now controllable
when processing csv files.