Changes for testability
For testing purposes, it's important to be able to process a single file
at a time, so I modified get_files so that if the 'range' argument looks
like a filename, the single file will be returned and the range logic
will be bypassed.

I also do a lot of testing and development without an internet connection,
so I added exception handling to treat failure to connect to GENUKI the
same way we treat a failure to resolve a GENUKI placename.