Updated to include List of Pieces for District and then list Unique names for Online Pieces.
Updated to include List of Pieces for District and then for Online Pieces to list Unique names.
Note: this version only 'recognises' fc2 Pieces as Online as we had discussed the possibility of setting the fc2 Piece Status to Online when the relevant VLD had been incorporated, so didnt want to write a lot of code that could then be dosguarded.