Updated to handle Place name changes
The addition of years with online data has been moved to the freecen2_contents calculate method as  previously it relied on freecen2_place.cen_data_years array. This correction was made because the cen_data_years array could include data for yers which have not yet been captured in freecen2_contents.
Freecen2_contents now recognises if a place_name has been updated and marks the county for a refresh.