Update of the validation process step 2
If Verbatim Place of Birth is not the same as the Alternative Place of Birth and records is valid then the Propagate POB action is available.
If Notes field is not null, then the Propagate Notes action is available.
Added adjustment of total warnings for the CSV file if warnings are accepted.