Update freecen_csv_entry.rb
Added Sequence_in_household = 1 to method that looks for the address associated with a dwelling for CSVproc data Search Results. In most cases the existing find would bring back the records in sequence_in_household order by luck(??. Adding this condition ensures that the record it gets is for the first individual for a dwelling which is where the address details are stored.